+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `ContextAwareHandler` interface and `MakeContextAware` function
+
 
 ## [0.21.1] - 2019-07-19